리뷰 작성 완료시 다이어로그 추가, 일부 워딩 수정

+++ app/src/main/java/com/example/kusithms_hdmedi_project/model/HospitalListResponse.kt
@@ -17,6 +17,6 @@ data class Hospital(
     val address: String = "",
     val hospitalId: Int = 0,
     val name: String = "",
-    val averageRating: Int = 0,
+    val averageRating: Double = 0.0,
     val numberOfReview: Int = 0
 )
